requirements.txt
Pillow
opencv-python
numpy
rich
